requirements.txt
fastapi
uvicorn
onnxruntime
tensorflow-cpu
opencv-python-headless
pillow
python-multipart
numpy
